requirements.txt
# AI/LLM
google-generativeai==0.3.0
langchain==0.3.0

# GitHub
PyGithub==2.4.0
requests==2.32.3

# Code Analysis
bandit==1.7.9
pylint==3.2.0
black==24.8.0
radon==6.0.1

# Testing
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2

# Database (for memory)
neo4j==5.15.0

# Logging
structlog==24.4.0